Fix a type, an ORM identity gap, and a misleading count
Three follow-up review points, all of which were real.

Annotation.scaleValue goes back to Integer. An annotator picks a whole number on a scale; only the value agreed at resolution may be an average, so Resolution.scaleValue stays a Double.

TaxonomySettings gains its own id and projectId. ORMLite has no equivalent of JPA's @Embedded, so it can only persist a type that has its own identity; without that the scale bounds could not be tied back to a project.

countSubmittedByAnnotator claimed earnings were derived from it, which was wrong in the same way countByRole was: earnings also exclude retired items and depend on the project being complete, so they are computed in EarningsService. The comment now says what the count is actually for.

+++ src/main/java/arbiter/model/annotation/Annotation.java
@@ -25,8 +25,11 @@ public class Annotation {
     /** Chosen label, for a taxonomy whose kind is SINGLE. */
     private Long labelId;
 
-    /** Numeric answer, for a taxonomy whose kind is SCALE. */
-    private Double scaleValue;
+    /**
+     * Numeric answer, for a taxonomy whose kind is SCALE. An annotator picks a whole number, so
+     * this is an int; only the value agreed at resolution may be an average.
+     */
+    private Integer scaleValue;
 
     /** The annotator's explanation. */
     private String rationale;
@@ -88,12 +91,12 @@ public class Annotation {
         }
     }
 
-    public Double getScaleValue() {
+    public Integer getScaleValue() {
         return scaleValue;
     }
 
     /** Sets the numeric answer and clears any chosen label, so only one answer shape is ever set. */
-    public void setScaleValue(Double scaleValue) {
+    public void setScaleValue(Integer scaleValue) {
         this.scaleValue = scaleValue;
         if (scaleValue != null) {
             this.labelId = null;
